Add Batchwriter. (#1)
Initial commit to harbourbridge repository with license/contrib files,
an initial README (mostly empty), and the first chunk of code: batchwriter.

Batchwriter is an abstraction on top of the standard Spanner Client
library that batches mutations together and manages the writing of
those mutations to Spanner. HarbourBridge uses this to efficiently
write data to Spanner.