Add tablemap. (#10)
Tablemap maintains the mappings between PostgreSQL tables and columns
and Spanner tables and columns. Note that PostgreSQL table and column
names can be essentially any string, but Spanner table and column
names must use a limited character set. So when we convert PostgreSQL
to Spanner, we potentially have to remap PostgreSQL table and column
names to make them legal Spanner.