Refactor step 2: remove uses of postgres and pg.
This commit:
- removes all uses of postgres and pg in variables and function names
  (mostly they are replaced with src).
- moves internal/postgres to internal (since most of that code will
  be generic code).

Note that while the vast majority of references are gone, there are
still some references to postgres in the code. These are either in
postgres specific code (e.g. in statement.go), or are in comments that
mention particular behaviour of postgres (leaving these in place for
the moment).