#87: feat: support NUMERIC (roll forward) (#104)
Fixes #87. Add support for Spanner NUMERIC to schema and data conversion (roll forward).

This was first submitted as PR98, but was rolled back because of client issues with big.Rat.

This PR consists of:
1) a clean revert of #102 (revert of #87)
2) a fix for the client issue (numeric values now passed as a string via spanner.NumericString)
3) additional integration tests (includes all core types e.g. numeric)
4) updates github workflow to use recent version of emulator (to get numeric support)

Original PR description:
Fixes #87. Add support for Spanner NUMERIC to schema and data conversion.

Before this PR, HarbourBridge mapped PostgreSQL and MySQL NUMERIC/DECIMAL types to Spanner's FLOAT64 because Spanner did not originally support NUMERIC. However, Spanner has started supporting NUMERIC type from v1.10.0 (around Sept/2020). We've been planning to support it in HarbourBridge since then.

We'd be holding off on doing so while all of the work for user-guided schema migration was ongoing (aka webui work). We've now reached a point where most of the potentially conflicting changes from that work have landed. So it's time to push ahead with NUMERIC support.