#136: fix: Handle non-boolean values for boolean columns. (#138)
This PR changes the behaviour during data conversion for boolean columns with non-boolean values: instead of generating an error, we now try to convert an int to a bool.

Old behaviour: HarbourBridge generates an error and rejects rows when boolean columns have non-boolean values.

New behaviour: We try to parse the value as an int64. If that succeeds, then we convert the int to a bool (0 maps to false, everything else maps to true), and we log an "unexpected" event detailing the conversion. If the conversion to int64 fails, the behavior is the same as current behaviour (we log an error and drop the row).

Closes #136