Map MySQL `AUTO_INCREMENT` columns to Spanner `IDENTITY` columns (#1227)
* Map auto increment source columns to Spanner IDENTITY columns

Note, I'm not using the AUTO_INCREMENT keyword or SERIAL type here because those don't allow setting of the skip range
or start counter with values. Additionally, they both alias to the IDENTITY clause I'm using here anyways, so it's
effectively the same, but this is more flexible.

* Remove sequence creation for AUTO_INCREMENT columns

The use of a sequence will no longer be needed. The next commit will update the AST printing to use the new
AUTO_INCREMENT keyword (Spanner SQL) and SERIAL type (PGSQL).

* Add support for migrating AUTO_INCREMENT columns from mysql dump files

The AST mapping will be added in a subsequent commit.

* Minor cleanup to align cassandra's GetColumnAutoGen impl with other DBs that don't support migration of auto incrementing columns

* Add note to set skip range or start counter with to avoid collisions for identity columns

* Add Identity as an option for Auto-Gen columns in the web UI

* Allow for setting of the skip range and start counter with values via the web UI

* Update docs

* Fix notation for identity column type (needed to use underscores instead of spaces...)

* Fix lint and integration test workflows

* Allow Go version 1.24 for lint tool (latest staticcheck requires at least 1.24)

---------

Co-authored-by: Vardhan Vinay Thigle <39047439+VardhanThigle@users.noreply.github.com>